+++ workflows/OCORaman/scripts/06_run_solar_abs_761_764_cpu.jl
@@ -80,7 +80,7 @@ geometry:
   sza: 19.0
   vza: [0.0]
   vaz: [0.0]
-  obs_alt: 0
+  obs_alt: [0]
 
 atmospheric_profile:
   T: $(yaml_array(T_PROFILE))
@@ -133,7 +133,7 @@ geometry:
   sza: 19.0
   vza: [0.0]
   vaz: [0.0]
-  obs_alt: 0
+  obs_alt: [0]
 
 atmospheric_profile:
   T: $(yaml_array(T_PROFILE))